fix: 修复lwIP在FreeBSD平台上的编译错误
应用上游lwIP 6df0ccc提交的修复：
- bridgeif_fdb.c: 移动#if LWIP_TIMERS指令位置
- perf.c: 修复fprintf格式字符串使用%lld而非%lu
- pcapif.c: 添加FreeBSD条件编译排除

+++ third-part/lwip/src/ports/unix/lib/perf.c
@@ -55,8 +55,8 @@ perf_print(unsigned long c1l, unsigned long c1h,
 void
 perf_print_times(struct tms *start, struct tms *end, char *key)
 {
-  fprintf(f, "%s: %lu\n", key, end->tms_stime - start->tms_stime);
-  fflush(NULL);  
+  fprintf(f, "%s: %lld\n", key, (long long)(end->tms_stime - start->tms_stime));
+  fflush(NULL);
 }
 
 void
